requirements.txt
numpy
pandas
sklearn
flask
flask-cors
gunicorn
PyYAML
dill
pymongo[srv]
dnspython
xgboost
optuna
evidently
-e.